Created cultural_genetic_algo.py
Implementation of a hexadoku solver which employs a cultural genetic algorithm. The algorithm generates a population of boards or 'individuals' which are filled with random values, then progressively improves the population by selecting individuals to reproduce, then replacing the individuals with the worst fitness values with the new offspring. Crossover and mutation is implemented for each new offspring. This process is repeated until an individual contains a valid solution to the puzzle.